Refactor while let on iterator lint to its module; rename for loop explicit counter to explicit counter loop
