+++ clippy_lints/src/loops/explicit_counter_loop.rs
